gh-700: Move to correct folder
