+++ database/utils/helper_functions.go
