+++ database/utils/helper_objects.go
