+++ restapi/configure_weaviate.go
@@ -44,10 +44,10 @@ import (
 	weaviateBroker "github.com/creativesoftwarefdn/weaviate/broker"
 	"github.com/creativesoftwarefdn/weaviate/config"
 	"github.com/creativesoftwarefdn/weaviate/database"
-	dblisting "github.com/creativesoftwarefdn/weaviate/database/connectors/listing"
-	connutils "github.com/creativesoftwarefdn/weaviate/database/connectors/utils"
+	dblisting "github.com/creativesoftwarefdn/weaviate/database/listing"
 	"github.com/creativesoftwarefdn/weaviate/database/schema"
 	db_local_schema_manager "github.com/creativesoftwarefdn/weaviate/database/schema_manager/local"
+	connutils "github.com/creativesoftwarefdn/weaviate/database/utils"
 	"github.com/creativesoftwarefdn/weaviate/graphqlapi"
 	graphqlnetwork "github.com/creativesoftwarefdn/weaviate/graphqlapi/network"
 	"github.com/creativesoftwarefdn/weaviate/lib/delayed_unlock"
