+++ restapi/handlers_things.go
@@ -16,7 +16,7 @@ import (
 	"fmt"
 
 	weaviateBroker "github.com/creativesoftwarefdn/weaviate/broker"
-	connutils "github.com/creativesoftwarefdn/weaviate/database/connectors/utils"
+	connutils "github.com/creativesoftwarefdn/weaviate/database/utils"
 	"github.com/creativesoftwarefdn/weaviate/database/schema"
 	"github.com/creativesoftwarefdn/weaviate/database/schema/kind"
 	"github.com/creativesoftwarefdn/weaviate/lib/delayed_unlock"
